Reads template resources as input-streams
When one tries to read a resource template from a jar via
io/resource, it returns a URL using the jar:file: protocol, which
isn't supported. This way, we return an input-stream, which is accepted
by io/copy.

This whole issue is masked when running from the REPL, since the
file can then be found on the filesystem, and relying on the jar
isn't necessary.